requirement.txt
asgiref==3.5.2
Django==4.1.3
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.3.0
Pygments==2.13.0
pytz==2022.6
sqlparse==0.4.3
